requirements.txt
Django==5.1.5
channels==4.0.0
channels-redis==4.2.0
djangorestframework==3.14.0
django-filter==23.5
redis==5.0.1
asgiref==3.8.1
websockets==12.0
daphne==4.0.0 